+++ GameShop/Controllers/WishlistController.cs
@@ -13,20 +13,27 @@ namespace GameShop.Controllers
     public class WishlistController : ControllerBase
     {
         private readonly IWishlistService _wishlistService;
+
         public WishlistController(IWishlistService wishlistService)
         {
             _wishlistService = wishlistService;
         }
+
         [HttpPost("UserID")]
         public async Task<IActionResult> AddWishlist(string UserID, AddWishlistRequest addWishlistRequest)
         {
-            var result =await _wishlistService.AddWishlist(UserID, addWishlistRequest);
+            if (!ModelState.IsValid)
+            {
+                return BadRequest();
+            }
+            var result = await _wishlistService.AddWishlist(UserID, addWishlistRequest);
             if (!result.IsSuccess)
             {
                 return BadRequest(result);
             }
             return Ok(result);
         }
+
         [HttpGet("UserID")]
         public async Task<IActionResult> GetWishlist(string UserID)
         {
@@ -40,9 +47,14 @@ namespace GameShop.Controllers
                 return Ok(result);
             }
         }
+
         [HttpDelete("UserID")]
         public async Task<IActionResult> DeleteItem(string UserID, [FromBody] DeleteItemRequest deleteItemRequest)
         {
+            if (!ModelState.IsValid)
+            {
+                return BadRequest();
+            }
             var result = await _wishlistService.DeleteItem(UserID, deleteItemRequest);
             if (!result.IsSuccess)
             {
@@ -54,5 +66,4 @@ namespace GameShop.Controllers
             }
         }
     }
-   
 }
